dumpcab.py: decompress more than just the first lzx block
With this, SETUP.EX_ seems to be complete and correct
(except for E8 jump translation).

This wraps the main loop in a "while more blocks" loop,
so look at this change with `git diff -w` to ignore the
reindentation (or add "?w=1" to the github url of the
commit).

Changes:
* Move outfile, trees, window up so that they're shared
  across lzx blocks
* Interpret pretree codes relative to existing tree values
* Fix off-by-one in output() when writing past end of window
* Do CFDATA block flushing based on total uncompressed bytes
  written, not based on uncompressed bytes from current
  lzx block