cam: use URL.createObjectURL() instead of reader.readAsDataURL
This makes "open image in new tab" work.

I will want to use readAsArrayBuffer() to look at EXIF data
later on, but calling createObjectURL() on the file instead of
on a blog in the readAsArrayBuffer() callback seems more direct
anyways, and I'll want to keep using that for the generated img
element even once I call readAsArrayBuffer().