res: Make output for empty VALUEs in VERSIONINFO match rc.exe
Empty strings after a VALUE's key followed by non-empty strings have
strange behavior: They cause the key to be written without padding,
the value data is written into the padding. This is a case that
hopefully doesn't happen in practice, and it kind of looks like
invalid .res output. In practice, this hopefully doesn't happen.

If the value is just an empty string and nothing else, then
having the empty string be completely ignored does the right
thing and kind of makes sense, so let's just do that for now.

See https://crbug.com/1473311