add web/map.html
A demo for a map that pans on trackpad pan gestures and
zooms for trackpad zoom gestures (two-finger pinch) on macOS.
Seems to work in Chrome, Firefox, Safari.

Don't know if it does the right thing on Windows or mobile,
probably not yet.

(Written by Gemini, mostly putting here so I can try it on
other devices using github.io.)